Rename files in magma backend for consistency with other libCEED backends
